+++ ConfigManager.cs
@@ -17,6 +17,12 @@ internal class ConfigManager
     internal ConfigEntry<int> minPlayersForDynamicCredits;
     internal ConfigEntry<int> maxPlayersForDynamicCredits;
 
+    internal ConfigEntry<int> dynamicStartCreditMinIncrease;
+    internal ConfigEntry<int> dynamicStartCreditMaxIncrease;
+
+    internal ConfigEntry<int> dynamicStartCreditAdditiveChange;
+    internal ConfigEntry<float> dynamicStartCreditMultiplicativeChange;
+
     internal ConfigEntry<bool> resetOnFirstDayUponReHost;
 
     internal ConfigEntry<bool> terminalCommandsEnabled;
@@ -28,6 +34,7 @@ internal class ConfigManager
     private const string generalSection = "General";
     private const string staticStartCreditsSection = "Static Start Credits";
     private const string dynamicStartCreditsSection = "Dynamic Start Credits";
+    private const string changingDynamicStartCreditsSection = "Changing Dynamic Start Credits";
     private const string randomStartCreditsSection = "Random Start Credits";
     private const string terminalCommandsSection = "Terminal Commands";
 
@@ -38,6 +45,10 @@ internal class ConfigManager
     private const string maxDynamicCreditPlayersKey = "Max Players";
     private const string minRandomCreditsKey = "Min Additional Start Credits";
     private const string maxRandomCreditsKey = "Max Additional Start Credits";
+    private const string dynamicStartCreditMinIncreaseKey = "Min Increase";
+    private const string dynamicStartCreditMaxIncreaseKey = "Max Increase";
+    private const string dynamicStartCreditAdditiveChangeKey = "Additive Change";
+    private const string dynamicStartCreditMultiplicativeChangeKey = "Multiplicative Change";
     private const string resetOnFirstDayUponReHostKey = "Reset on First Day Upon Re-Host";
     private const string terminalCommandPrefixKey = "Prefix For All Commands";
     private const string terminalCommandHelpKey = "Help Command Keyword";
@@ -52,9 +63,17 @@ internal class ConfigManager
 
     private readonly string dynamicStartCreditIncreasePerPlayerDescription = "Gives extra credits when someone joins the game for the first time and the crew hasn't landed yet.\n\nHost is included.\n\nThis is based on player count rather than WHO joins. If 2 people join, then it adds the bonus twice and until there are more than 3 people (host + 2), it won't apply the bonus again.";
 
-    private readonly string minPlayersForDynamicCreditsDescription = "How many players have to be present in the lobby for dynamic credits to be applied.\n\nEx. 0 means that host counts and this will be applied to the host and any other who joins.\n\nEx. 1 means that someone else other than the host would need to join before we apply the \"Start Credit Increase Per Player\" increase.";
+    private readonly string minPlayersForDynamicCreditsDescription = "How many players have to be present in the lobby for dynamic credits to be applied.\n\nEx. 0 means that host counts and this will be applied to the host and any other who joins.\n\nEx. 2 means that 2 players other than the host would need to join before we apply the \"Start Credit Increase Per Player\" increase at all. This would also mean that the counting starts from the third player with host being ignored meaning the bonus is applied ONCE in this scenario.";
 
     private readonly string maxPlayersForDynamicCreditsDescription = "How many players can be in the lobby before we stop applying the \"Start Credit Increase Per Player\" increase.\n\nEx. 4 would mean that the bonus is applied 4 times, host included. If a 5th person joined, host included, then the bonus would NOT be applied.\n\n-1 Means there is no maximum cap.";
+
+    private readonly string dynamicStartCreditMinIncreaseDescription = "Minimum start credit increase value per player.\n\nUsed for when reducing/increasing player based start credits.\n\nIf this is less than 0, the starting credits could go DOWN if too many join.";
+
+    private readonly string dynamicStartCreditMaxIncreaseDescription = "Maximum start credit increase value per player.\n\nUsed for when reducing/increasing player based start credits.\n\n-1 means there is no maximum cap.";
+
+    private readonly string dynamicStartCreditAdditiveChangeDescription = "When someone joins, this is added to the Dynamic Start Credits per player.\n\nThis is applied BEFORE the multiplicative change.\n\nEx. Dynamic Start Credits = 100. Additive Change = -10. Multiplicative Change = 0.5. 1 players join. First the credits are 100 for the first player, aka. host. Then the Dynamic Start Credits are applied again for the second player. (100 + (-10)) * 0.5 = 45. The next credit increase would be 45 and we would have 145 credits with host and 1 other.";
+
+    private readonly string dynamicStartCreditMultiplicativeChangeDescription = "When someone joins, Dynamic Start Credits are multiplied by this PER PLAYER.\n\nThis is applied BEFORE the multiplicative change.\n\nEx. Dynamic Start Credits = 100. Additive Change = -10. Multiplicative Change = 0.5. 2 players join. First the credits are 100 for the first player, aka. host. Then the Dynamic Start Credits are applied again for the second player. (100 + (-10)) * 0.5 = 45. The next credit increase would be 45 and we would have 145 credits with host and 1 other. Then the Dynamic Start Credits are applied again for the third player. (45 + (-10)) * 0.5) = 17.5. Now we have 162.5 credits with host and 2 others.";
 
     private readonly string randomCreditsEnabledDescription = "If true, the crew will get a random amount of credits.\n\nThis will be applied IN ADDITION to any other credits like \"Static Start Credits\".\n\nIf you just want any random amount, you should set \"Static Start Credits\" to 0.\n\nEx. if \"Static Start Credits\" is 200 and this is in range of -100 to 150, then the start credits could be from 100 (200 - 100) to 350 (200 + 150).";
 
@@ -64,9 +83,9 @@ internal class ConfigManager
 
     private readonly string resetOnFirstDayUponReHostDescription = "When you create a lobby (or get fired), you can buy items and then save and exit. When you re-host, the bought items are not saved but you have still lost the credits. This is a \"fix\" for that.\n\nIf this is enabled, you will get your start credits back if you re-host the game on your first day (day 0).\n\nThis is disabled by default because some mods still save purchases on day 0 so this setting could be \"abused\" in that sense.";
 
-    private readonly string terminalCommandsEnabledDescription = "If any of the terminal commands from this mod are enabled. \n\nCurrent terminal commands are:\n\"reload\"\n\"help\"\n\nSpaces are NOT allowed.\n\nYou can disable individual commands by leaving them empty. It is not recommended but you can leave the prefix empty and still run commands.";
+    private readonly string terminalCommandsEnabledDescription = "If any of the terminal commands from this mod are enabled. \n\nCurrent terminal commands are:\n\"reload\"\n\"help\"\n\nSpaces are NOT allowed.\n\nYou can disable individual commands by leaving them empty.";
 
-    private readonly string terminalCommandPrefixDescription = "Every terminal command from this mod needs to have this keyword before them. Case doesn't matter.\n\nEx. if this setting is StartCreditsPlus, you can reload credits by typing \"STARTcreditsPLUS REload\" or \"StartCreditsPlus reload\", etc.";
+    private readonly string terminalCommandPrefixDescription = "Every terminal command from this mod needs to have this keyword before them. Case doesn't matter.\n\nEx. if this setting is StartCreditsPlus, you can reload credits by typing \"STARTcreditsPLUS REload\" or \"StartCreditsPlus reload\", etc.\n\nYou can leave the prefix empty and still run commands but it isn't recommended due to possible overlapping commands.";
 
     private readonly string terminalCommandHelpDescription = "Shows all available terminal commands and their descriptions.\n\nThis is useful if you forget what a command does or if you want to see all available commands.";
 
@@ -94,6 +113,12 @@ internal class ConfigManager
         dynamicStartCreditIncreasePerPlayer = config.Bind(dynamicStartCreditsSection, startCreditIncreasePerPlayerKey, 15, dynamicStartCreditIncreasePerPlayerDescription);
         minPlayersForDynamicCredits = config.Bind(dynamicStartCreditsSection, minDynamicCreditPlayersKey, 0, minPlayersForDynamicCreditsDescription);
         maxPlayersForDynamicCredits = config.Bind(dynamicStartCreditsSection, maxDynamicCreditPlayersKey, -1, maxPlayersForDynamicCreditsDescription);
+
+        dynamicStartCreditMinIncrease = config.Bind(changingDynamicStartCreditsSection, dynamicStartCreditMinIncreaseKey, 0, dynamicStartCreditMinIncreaseDescription);
+        dynamicStartCreditMaxIncrease = config.Bind(changingDynamicStartCreditsSection, dynamicStartCreditMaxIncreaseKey, -1, dynamicStartCreditMaxIncreaseDescription);
+
+        dynamicStartCreditAdditiveChange = config.Bind(changingDynamicStartCreditsSection, dynamicStartCreditAdditiveChangeKey, -5, dynamicStartCreditAdditiveChangeDescription);
+        dynamicStartCreditMultiplicativeChange = config.Bind(changingDynamicStartCreditsSection, dynamicStartCreditMultiplicativeChangeKey, 1f, dynamicStartCreditMultiplicativeChangeDescription);
 
         // Terminal commands
         terminalCommandsEnabled = config.Bind(terminalCommandsSection, enabledKey, true, terminalCommandsEnabledDescription);
@@ -145,6 +170,18 @@ internal class ConfigManager
                     break;
                 case maxDynamicCreditPlayersKey:
                     maxPlayersForDynamicCredits.Value = (int)args.ChangedSetting.BoxedValue;
+                    break;
+                case dynamicStartCreditMinIncreaseKey:
+                    dynamicStartCreditMinIncrease.Value = (int)args.ChangedSetting.BoxedValue;
+                    break;
+                case dynamicStartCreditMaxIncreaseKey:
+                    dynamicStartCreditMaxIncrease.Value = (int)args.ChangedSetting.BoxedValue;
+                    break;
+                case dynamicStartCreditAdditiveChangeKey:
+                    dynamicStartCreditAdditiveChange.Value = (int)args.ChangedSetting.BoxedValue;
+                    break;
+                case dynamicStartCreditMultiplicativeChangeKey:
+                    dynamicStartCreditMultiplicativeChange.Value = (float)args.ChangedSetting.BoxedValue;
                     break;
             }
         }